+++ src/toolbox_continu_inzicht/base/adapters/output/python.py
@@ -1,0 +1,11 @@
+import pandas as pd
+
+
+def output_python(output_config: dict, df: pd.DataFrame) -> None:
+    """Wrapper voor de functionalitiet om dataframes vanuit python te ondersteunen
+
+    Returns:
+    --------
+    None
+    """
+    pass
