--- tests/src/base/__init__.py
@@ -1,1 +1,0 @@
-"""Tests for continu_inzicht_toolbox.base module"""
