+++ tests/src/base/data_sets/test_load_external_dataset_config.yaml
@@ -1,0 +1,31 @@
+GlobalVariables:
+    rootdir: 'tests/src/base/data_sets'
+    calc_time: '2024-11-18 08:00:00'
+    input_plugin_path: "tests/src/base/data_sets"
+    output_plugin_path: "tests/src/base/data_sets"
+
+DataAdapter:
+    default_options:
+        postgresql_database:
+            database: 'citoolbox'
+            schema: 'citoolbox_schema'
+        csv:
+            sep: ","
+
+    MyCSV_in: 
+        type: csv
+        path: 'test_csv_in.csv'
+    MyCSV_out: 
+        type: csv
+        path: 'test_csv_out.csv'
+    MyPostgresql: 
+        type: postgresql_database
+        table: data_test
+    MyNetCDF_in:
+        type: netcdf
+        path: 'test_netcdf_in.nc'
+    MyNetCDF_out:
+        type: netcdf
+        path: 'test_netcdf_out.nc'
+    my_df_python:
+        type: python
